Rename plugin to loaderplugin
- As that is what the module is, and also to match with upstream.
- The registry name should also be renamed, but the impact is not yet
  known.  Will only be done if a deprecation mechanics is in place.
