+++ src/calmjs/rjs/loaderplugin.py
